+++ R/summarize-growth.R
@@ -12,16 +12,25 @@
 #' The fitness proxies returned are the parameters of the logistic equation
 #' and the area under the curve (a measure that integrates the effects
 #' of \eqn{N_0}, K, and r). See \code{\link{gcfit}} for more documentation on these.
-#' @param data_t    A vector of timepoints (data_n must also
-#'                  be provided and be the same length).
-#' @param data_n    A vector of cell counts or absorbance readings.
-#' @param t_trim    Measurements taken after this time should not be included
-#'                  in fitting the curve. If stationary phase is variable,
-#'                  this may give you a better fit. A value of 0 means no
-#'                  trimming. Defaults to no trimming (0).
-#' @return          An object of type gcfit containing the "fitness" proxies,
-#'                  as well as the input data and the fitted model.
-#'                  See \code{\link{gcfit}}
+#' @param data_t     A vector of timepoints (data_n must also
+#'                   be provided and be the same length).
+#' @param data_n     A vector of cell counts or absorbance readings.
+#' @param t_trim     Measurements taken after this time should not be included
+#'                   in fitting the curve. If stationary phase is variable,
+#'                   this may give you a better fit. A value of 0 means no
+#'                   trimming. Defaults to no trimming (0).
+#' @param bg_correct The background correction method to use. No background
+#'                   correction is performed for the default "none". Specifying
+#'                   "min" subtracts the smallest value in a column from all the
+#'                   rows in that column, and specifying "blank" subtracts
+#'                   the values from the blank vector from the data_n vector.
+#' @param blank      A vector of absorbance readings from a blank well
+#'                   (typically contains only media) used for background
+#'                   correction. The corresponding blank value is subtracted
+#'                   from the data_n vector for each timepoint. Defaults to NA.
+#' @return           An object of type gcfit containing the "fitness" proxies,
+#'                   as well as the input data and the fitted model.
+#'                   See \code{\link{gcfit}}
 #' @examples
 #' # We can check that the parameters that are found are the same
 #' # as we use to generate fake experimental data. To do so, let's first
@@ -53,7 +62,8 @@
 #' plot(gc)
 #'
 #' @export
-SummarizeGrowth <- function(data_t, data_n, t_trim = 0) {
+SummarizeGrowth <- function(data_t, data_n, t_trim = 0,
+                            bg_correct = "min", blank = NA) {
 
   # make sure that both inputs are vectors
   if (is.list(data_t) == TRUE) {
@@ -69,18 +79,58 @@ SummarizeGrowth <- function(data_t, data_n, t_trim = 0) {
   stopifnot(is.vector(data_t))
   stopifnot(is.vector(data_n))
 
+  # make sure that the inputs are valid
+  if (!is.numeric(data_t) |!is.numeric(data_n)) {
+    stop("Error: The input data (data_t and data_n) must be numeric.")
+  }
+  if (length(data_t) != length(data_n)) {
+    stop("Error: The input data (data_t and data_n) must have the same number of rows")
+  }
+
+  # make sure that the background correction method is valid
+  if (!bg_correct %in% c("none", "min", "blank")) {
+    stop(paste0(bg_correct, "is not a valid option for bg_correct"))
+  }
+
+  # check for correctness of the blank (background correction) vector
+  if (bg_correct == "blank") {
+    if (is.list(blank) == TRUE) {
+      tryCatch( {blank <- unlist(blank)},
+                error = function(e) {stop("blank is not a vector")}
+      )
+      stopifnot(is.vector(blank))
+
+      if (!is.numeric(blank)) {
+        stop("Error: The blank data must be numeric.")
+      }
+      if (length(blank) != length(data_n)) {
+        stop("Error: The input data (data_n) and the background correction data (blank) must have the same number of rows.")
+      }
+    }
+  }
+
   # check t_trim parameter and set dependent variables
   if (t_trim > 0) {
     t_max <- t_trim
     data_n <- data_n[data_t < t_trim]
     data_t <- data_t[data_t < t_trim]
+    if (bg_correct == "blank") {
+      blank <- blank[data_t < t_trim]
+    }
   } else {
     t_max <- max(data_t, na.rm=TRUE)
   }
 
+  #do the background correction
+  if (bg_correct == "blank") {
+    data_n <- data_n - blank
+    data_n[data_n < 0] <- 0    # ensure readings are at least 0
+  }
+  else if (bg_correct == "min") {
+    data_n <- data_n - min(data_n)
+  }
 
-  auc_e <- EmpiricalAreaUnderCurve(data_t, data_n, t_max)
-  log_mod <- FitLogistic(data_t, data_n)
+  log_mod <- FitLogistic(data_t, data_n )
 
   p <- summary(log_mod)$coefficients
   k <- p[1]
@@ -93,13 +143,15 @@ SummarizeGrowth <- function(data_t, data_n, t_trim = 0) {
   r_se <- p[6]
   r_p <- p[12]
 
-  # get the inflection point
+  # get the inflection point, DT, auc, sigma, df
   t_inflection <- TAtInflection(k, n0, r)
 
   DT <- MaxDt(r)
-  auc_l <- AreaUnderCurve(k, n0, r, 0, t_max)$value
   sigma <- summary(log_mod)$sigma
   df <- summary(log_mod)$df[2]
+
+  auc_l <- AreaUnderCurve(k, n0, r, 0, t_max)$value
+  auc_e <- EmpiricalAreaUnderCurve(data_t, data_n, t_max)
 
   vals <- c(k, k_se, k_p, n0, n0_se, n0_p,
             r, r_se, r_p, sigma, df,