+++ NEWS.md
@@ -1,3 +1,15 @@
+# growthcurver 0.2.0
+
+This is a minor release containing one new function to summarize the growth 
+curves for many wells with a single command. Background correction has also
+been added.
+
+* Added `SummarizeGrowthByPlate` 
+* Extended the `plot` method for `gc_fit` objects to take standard plot options
+* Added background correction (option `bg_correct`)
+* Updated the vignette to include the new options
+* Set the default `bg_correct` option to `min`
+
 # growthcurver 0.1.0
 
 * Initial release